add test for make_serial.i2c
